+++ crates/ox_content_napi/src/transform_bindings.rs
@@ -23,6 +23,7 @@ mod file_tree_options;
 mod keyboard_keys_options;
 mod magic_link_options;
 mod math_options;
+mod mdast;
 mod partials_options;
 mod steps_options;
 mod transform_options;
@@ -47,6 +48,7 @@ pub use file_tree_options::JsFileTreeOptions;
 pub use keyboard_keys_options::JsKeyboardKeysOptions;
 pub use magic_link_options::{JsMagicLinkAlias, JsMagicLinkImageOverride, JsMagicLinkOptions};
 pub use math_options::JsMathOptions;
+pub use mdast::{JsMdastTransformResult, transform_from_mdast, transform_mdast};
 pub use partials_options::JsPartialsOptions;
 pub use steps_options::JsStepsOptions;
 pub use transform_options::JsTransformOptions;
@@ -172,78 +174,6 @@ pub fn transform(source: String, options: Option<JsTransformOptions>) -> Transfo
             let opts = options.unwrap_or_default();
             let core_options = opts.into();
             MarkdownTransformer::from_options(&core_options).transform(&source).into()
-        },
-        || TransformResult {
-            html: String::new(),
-            frontmatter: "{}".to_string(),
-            toc: vec![],
-            errors: vec![crate::ffi::UNEXPECTED_PANIC.to_string()],
-            imports: vec![],
-            exports: vec![],
-            components: vec![],
-        },
-    )
-}
-
-/// A parsed document on its way to a JavaScript `transformers` hook.
-#[napi(object)]
-pub struct JsMdastTransformResult {
-    /// The tree, as mdast JSON.
-    pub ast_json: String,
-    /// Frontmatter as JSON, which the tree has no room for.
-    pub frontmatter: String,
-    /// Preprocessing and parse errors collected so far.
-    pub errors: Vec<String>,
-}
-
-/// Runs a transform up to the point where the tree exists.
-///
-/// The counterpart to `transformFromMdast`: frontmatter is parsed and the
-/// opt-in Markdown features are expanded, then the tree is handed over as
-/// JSON for a `transformers` hook to rewrite.
-#[napi(js_name = "transformMdast")]
-pub fn transform_mdast(
-    source: String,
-    options: Option<JsTransformOptions>,
-) -> JsMdastTransformResult {
-    crate::ffi::recover(
-        || {
-            let core_options = options.unwrap_or_default().into();
-            let result =
-                MarkdownTransformer::from_options(&core_options).transform_mdast_json(&source);
-            JsMdastTransformResult {
-                ast_json: result.ast_json,
-                frontmatter: result.frontmatter,
-                errors: result.errors,
-            }
-        },
-        || JsMdastTransformResult {
-            ast_json: String::new(),
-            frontmatter: "{}".to_string(),
-            errors: vec![crate::ffi::UNEXPECTED_PANIC.to_string()],
-        },
-    )
-}
-
-/// Finishes a transform from an mdast a JavaScript `transformers` hook may
-/// have rewritten.
-///
-/// `transformMdastRaw` produces the tree; this renders it and runs
-/// everything that follows rendering — HTML postprocessing, sanitization,
-/// the table of contents, and the MDX metadata — so a rewritten tree loses
-/// none of it. `frontmatterJson` is carried through untouched.
-#[napi(js_name = "transformFromMdast")]
-pub fn transform_from_mdast(
-    ast_json: String,
-    frontmatter_json: String,
-    options: Option<JsTransformOptions>,
-) -> TransformResult {
-    crate::ffi::recover(
-        || {
-            let core_options = options.unwrap_or_default().into();
-            MarkdownTransformer::from_options(&core_options)
-                .transform_from_mdast_json(&ast_json, &frontmatter_json)
-                .into()
         },
         || TransformResult {
             html: String::new(),
